refactor: split graphql code base
