--- src/routes/graphql/types/root.ts
@@ -1,176 +1,0 @@
-import { GraphQLList, GraphQLNonNull, GraphQLObjectType, GraphQLString } from 'graphql';
-import { User } from './user.js';
-import { UUIDType } from './uuid.js';
-import { Profile } from './profile.js';
-import { Post } from './post.js';
-import {
-  ChangePostInput,
-  ChangeProfileInput,
-  ChangeUserInput,
-  CreatePostInput,
-  CreateProfileInput,
-  CreateUserInput,
-} from './inputs.js';
-import { MemberType, MemberTypeId } from './membertype.js';
-
-export const MutationRootType = new GraphQLObjectType({
-  name: 'MutationRootType',
-  fields: {
-    createUser: {
-      type: new GraphQLNonNull(User),
-      args: {
-        dto: { type: new GraphQLNonNull(CreateUserInput) },
-      },
-      resolve: async (parent, args, context) => {
-        const { name, balance } = args;
-        const { prisma } = context;
-        return prisma.user.create({ data: { name, balance } });
-      },
-    },
-    createProfile: {
-      type: new GraphQLNonNull(Profile),
-      args: {
-        dto: { type: new GraphQLNonNull(CreateProfileInput) },
-      },
-      resolve: async () => {},
-    },
-    createPost: {
-      type: new GraphQLNonNull(Post),
-      args: {
-        dto: { type: new GraphQLNonNull(CreatePostInput) },
-      },
-      resolve: async () => {},
-    },
-    changePost: {
-      type: new GraphQLNonNull(Post),
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-        dto: { type: new GraphQLNonNull(ChangePostInput) },
-      },
-      resolve: async () => {},
-    },
-    changeProfile: {
-      type: new GraphQLNonNull(Profile),
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-        dto: { type: new GraphQLNonNull(ChangeProfileInput) },
-      },
-      resolve: async () => {},
-    },
-    changeUser: {
-      type: new GraphQLNonNull(User),
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-        dto: { type: new GraphQLNonNull(ChangeUserInput) },
-      },
-      resolve: async () => {},
-    },
-    deleteUser: {
-      type: new GraphQLNonNull(GraphQLString),
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {},
-    },
-    deletePost: {
-      type: new GraphQLNonNull(GraphQLString),
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {},
-    },
-    deleteProfile: {
-      type: new GraphQLNonNull(GraphQLString),
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {},
-    },
-    subscribeTo: {
-      type: new GraphQLNonNull(GraphQLString),
-      args: {
-        userId: { type: new GraphQLNonNull(UUIDType) },
-        authorId: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {},
-    },
-    unsubscribeFrom: {
-      type: new GraphQLNonNull(GraphQLString),
-      args: {
-        userId: { type: new GraphQLNonNull(UUIDType) },
-        authorId: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {},
-    },
-  },
-});
-
-export const QueryRootType = new GraphQLObjectType({
-  name: 'QueryRootType',
-  fields: {
-    memberTypes: {
-      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(MemberType))),
-
-      resolve: async () => {
-        return '';
-      },
-    },
-    memberType: {
-      type: MemberType,
-      args: {
-        id: { type: new GraphQLNonNull(MemberTypeId) },
-      },
-      resolve: async () => {
-        return '';
-      },
-    },
-    users: {
-      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(User))),
-
-      resolve: async () => {
-        return '';
-      },
-    },
-    user: {
-      type: User,
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {
-        return '';
-      },
-    },
-    posts: {
-      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(Post))),
-
-      resolve: async () => {
-        return '';
-      },
-    },
-    post: {
-      type: Post,
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {
-        return '';
-      },
-    },
-    profiles: {
-      type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(Profile))),
-
-      resolve: async () => {
-        return '';
-      },
-    },
-    profile: {
-      type: Profile,
-      args: {
-        id: { type: new GraphQLNonNull(UUIDType) },
-      },
-      resolve: async () => {
-        return '';
-      },
-    },
-  },
-});
